# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Referral & Commission Platform API
  version: 1.0.0-draft.1
  summary: Phase 0 draft of the v1 API (spec §20). Generated docs replace this file from Phase 2 on; until then this file is the contract.
  description: |
    Conventions (see docs/api.md):
    - All paths are under `/v1`; the gateway may expose them as `/api/v1`.
    - Money is integer minor units plus ISO-4217 currency. Never floats.
    - Timestamps are RFC 3339 in UTC.
    - List endpoints use cursor pagination: `?cursor=&limit=` (max 100), response `{items, next_cursor}`.
    - Successful responses return the resource directly; errors use the `Error` envelope.
    - `Idempotency-Key` is required on every non-event POST that moves money.
    - Service endpoints act only for the product bound to the API credential. A product is never taken from the request body.
  license:
    name: Proprietary
servers:
  - url: https://api.ref.example.com/v1
  - url: http://localhost:8000/v1
    description: local
...
security: []
paths:
...
  /public/products/{slug}/theme:
    get:
      tags: [public]
      operationId: getProductTheme
      summary: Portal theme for a product (§18, ADR 0019)
      parameters:
        - name: slug
          in: path
          required: true
          schema: { $ref: '#/components/schemas/ProductSlug' }
      responses:
        '200':
          description: Theme
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductTheme' }
        '404': { $ref: '#/components/responses/Error' }

  # ---------------------------------------------------------------- auth
  /auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Password login for direct partners (§5.3)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 1, maxLength: 256, writeOnly: true }
                totp_code: { type: string, pattern: '^[0-9]{6}$', description: Required for admin roles }
      responses:
        '200': { $ref: '#/components/responses/Session' }
        '401': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /auth/refresh:
    post:
      tags: [auth]
      operationId: refreshSession
      summary: Rotate the refresh token (HttpOnly cookie). Reuse of a rotated token revokes the whole family.
      parameters:
        - name: refresh_token
          in: cookie
          required: true
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/Session' }
        '401': { $ref: '#/components/responses/Error' }
  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      security: [ { userJwt: [] } ]
      responses:
        '204': { description: Logged out; refresh token family revoked }
  /auth/handoff:
    post:
      tags: [auth, service]
      operationId: createHandoff
      summary: Product backend requests a single-use SSO handoff token (60 s) for a product account (§5.3)
      security: [ { serviceKey: [ 'referrers:enroll' ] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [external_user_id]
              properties:
                external_user_id: { $ref: '#/components/schemas/ExternalId' }
                return_path: { type: string, pattern: '^/[A-Za-z0-9/_-]*$', maxLength: 128, description: Portal-relative path only }
      responses:
        '201':
          description: Handoff token
          content:
            application/json:
              schema:
                type: object
                required: [redirect_url, expires_at]
                properties:
                  redirect_url: { type: string, format: uri, description: 'portal/auth/callback?token=…' }
                  expires_at: { $ref: '#/components/schemas/Timestamp' }
        '404': { $ref: '#/components/responses/Error' }
  /auth/handoff/exchange:
    post:
      tags: [auth]
      operationId: exchangeHandoff
      summary: Portal exchanges a handoff token for a session (single use)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [token]
              properties:
                token: { type: string, maxLength: 512 }
      responses:
        '200': { $ref: '#/components/responses/Session' }
        '401': { $ref: '#/components/responses/Error' }

  # ---------------------------------------------------------------- service
...
  /credits/balance:
    get:
      tags: [service]
      operationId: getCreditBalance
      summary: Spendable credit for the referrer linked to a product account (§11)
      security: [ { serviceKey: [ 'credits:write' ] } ]
      parameters:
        - name: external_user_id
          in: query
          required: true
          schema: { $ref: '#/components/schemas/ExternalId' }
        - name: currency
          in: query
          required: true
          schema: { $ref: '#/components/schemas/Currency' }
      responses:
        '200':
          description: Balance
          content:
            application/json:
              schema:
                type: object
                required: [currency, spendable, max_per_invoice_pct]
                properties:
                  currency: { $ref: '#/components/schemas/Currency' }
                  spendable: { $ref: '#/components/schemas/MinorAmount' }
                  max_per_invoice_pct: { type: integer, minimum: 0, maximum: 100 }
        '404':
          description: CREDIT_ACCOUNT_NOT_LINKED (product account not linked to a referrer)
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /credits/reservations:
    post:
      tags: [service]
      operationId: reserveCredit
      summary: Reserve credit for an order (two-phase, TTL 15 min)
      security: [ { serviceKey: [ 'credits:write' ] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReservationRequest' }
      responses:
        '201':
          description: Reserved (reserved_amount may be less than requested, possibly 0)
          content: { application/json: { schema: { $ref: '#/components/schemas/Reservation' } } }
        '200':
          description: Idempotent replay
          content: { application/json: { schema: { $ref: '#/components/schemas/Reservation' } } }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /credits/reservations/{reservation_id}/capture:
    post:
      tags: [service]
      operationId: captureCredit
      summary: Capture a reservation after the product's charge succeeds (ADR 0011). Idempotent.
      security: [ { serviceKey: [ 'credits:write' ] } ]
      parameters:
        - $ref: '#/components/parameters/ReservationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [applied_amount, payment_id]
              properties:
                applied_amount: { $ref: '#/components/schemas/MinorAmount' }
                payment_id: { $ref: '#/components/schemas/ExternalId' }
      responses:
        '200':
          description: Captured (state CAPTURED, PARTIALLY_CAPTURED or LATE_CAPTURED)
          content: { application/json: { schema: { $ref: '#/components/schemas/Reservation' } } }
        '404': { $ref: '#/components/responses/Error' }
  /credits/reservations/{reservation_id}/release:
    post:
      tags: [service]
      operationId: releaseCredit
      summary: Release a reservation (payment failed or abandoned). Idempotent.
      security: [ { serviceKey: [ 'credits:write' ] } ]
      parameters:
        - $ref: '#/components/parameters/ReservationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Released
          content: { application/json: { schema: { $ref: '#/components/schemas/Reservation' } } }
        '404': { $ref: '#/components/responses/Error' }
        '409':
          description: RESERVATION_ALREADY_CAPTURED
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /credits/refunds:
    post:
      tags: [service]
      operationId: refundCredit
      summary: Return captured credit to the referrer after the product refunds a payment that used it (§12)
      security: [ { serviceKey: [ 'credits:write' ] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reservation_id, amount, currency, idempotency_key]
              properties:
                reservation_id: { type: string, format: uuid }
                refund_id: { $ref: '#/components/schemas/ExternalId' }
                amount: { $ref: '#/components/schemas/PositiveMinorAmount' }
                currency: { $ref: '#/components/schemas/Currency' }
                idempotency_key: { $ref: '#/components/schemas/IdempotencyKeyValue' }
      responses:
        '201':
          description: Credit returned
          content:
            application/json:
              schema:
                type: object
                required: [credit_refund_id, amount, currency, remaining_refundable]
                properties:
                  credit_refund_id: { type: string, format: uuid }
                  amount: { $ref: '#/components/schemas/MinorAmount' }
                  currency: { $ref: '#/components/schemas/Currency' }
                  remaining_refundable: { $ref: '#/components/schemas/MinorAmount' }
        '200': { description: Idempotent replay }
        '422':
          description: CREDIT_REFUND_EXCEEDS_CAPTURED
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
components:
  securitySchemes:
    serviceKey:
      type: http
      scheme: bearer
      bearerFormat: '<key_id>.<secret>'
      description: Product API client credential with scopes. Never sent from a browser.
    eventSignature:
      type: apiKey
      in: header
      name: X-Signature
      description: HMAC-SHA256 over `timestamp + "." + raw_body` with the client's signing secret.
    userJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 15-minute access token held in memory by the portal/admin.
  parameters:
...
    ReservationId:
      name: reservation_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { $ref: '#/components/schemas/IdempotencyKeyValue' }
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Too many requests
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Session:
      description: Session established. Refresh token set as HttpOnly, Secure, SameSite=Lax cookie.
      headers:
        Set-Cookie: { schema: { type: string } }
      content:
        application/json:
          schema:
            type: object
            required: [access_token, expires_in, user]
            properties:
              access_token: { type: string }
              token_type: { type: string, const: Bearer }
              expires_in: { type: integer, const: 900 }
              user: { $ref: '#/components/schemas/Me' }
  schemas:
    Error:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, const: false }
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string, pattern: '^[A-Z][A-Z0-9_]+$', description: See the error registry in docs/api.md }
            message: { type: string }
            request_id: { type: string }
            details: { type: object }
    Timestamp: { type: string, format: date-time }
    Currency: { type: string, pattern: '^[A-Z]{3}$' }
    MinorAmount: { type: integer, minimum: 0, format: int64 }
...
    PositiveMinorAmount: { type: integer, minimum: 1, format: int64 }
    ExternalId: { type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9._:-]+$' }
    ProductSlug: { type: string, pattern: '^[a-z][a-z0-9-]{1,31}$' }
    IdempotencyKeyValue: { type: string, minLength: 8, maxLength: 128, pattern: '^[A-Za-z0-9._:-]+$' }
...
    ProductTheme:
      type: object
      properties:
        product: { $ref: '#/components/schemas/ProductSlug' }
        name: { type: string }
        logo_url: { type: string, format: uri }
        colors:
          type: object
          additionalProperties: { type: string, pattern: '^#[0-9a-fA-F]{6}$' }
        font_family: { type: string }
...
    ReservationRequest:
      type: object
      additionalProperties: false
      required: [external_user_id, currency, requested_amount, order_ref, idempotency_key]
      properties:
        external_user_id: { $ref: '#/components/schemas/ExternalId' }
        currency: { $ref: '#/components/schemas/Currency' }
        requested_amount: { $ref: '#/components/schemas/PositiveMinorAmount' }
        invoice_amount: { $ref: '#/components/schemas/PositiveMinorAmount', description: Required to apply the per-invoice cap; defaults to requested_amount }
        order_ref: { type: string, maxLength: 128 }
        idempotency_key: { $ref: '#/components/schemas/IdempotencyKeyValue' }
    Reservation:
      type: object
      required: [reservation_id, state, currency, reserved_amount, expires_at]
      properties:
        reservation_id: { type: string, format: uuid }
        state: { type: string, enum: [RESERVED, CAPTURED, PARTIALLY_CAPTURED, RELEASED, EXPIRED, LATE_CAPTURED] }
        currency: { $ref: '#/components/schemas/Currency' }
        requested_amount: { $ref: '#/components/schemas/MinorAmount' }
        reserved_amount: { $ref: '#/components/schemas/MinorAmount' }
        captured_amount: { $ref: '#/components/schemas/MinorAmount' }
        expires_at: { $ref: '#/components/schemas/Timestamp' }
    Me:
      type: object
      required: [id, roles]
      properties:
        id: { type: string, format: uuid }
        display_name: { type: string }
        email: { type: [string, 'null'] }
        email_verified: { type: boolean }
        phone_verified: { type: boolean }
        kyc_status: { type: string, enum: [NONE, PENDING, VERIFIED, REJECTED] }
        roles: { type: array, items: { type: string } }
        locale: { type: string, enum: [en, bn] }
        time_zone: { type: string }
...